+++ R/clust_it.R
@@ -19,7 +19,7 @@
 clust_it <- function(pop, m = 5, l = 4, g1 = 5, g2 = g1 * 0.5, d1 = 1000, d2 = d1 * 2){
         clusterizados <- as.list(NA)
         j <- 1 #iteration index
-        n <- 1 #pob remainder
+        n <- nrow(pop) #pob remainder
         while (j <= m & n >=  1 & ceiling(nrow(pop) / (ifelse(j <= l, g1, g2) * 0.75)) >= 2) { 
                 pop <- clust_pop(pop, ceiling(nrow(pop) / (ifelse(j <= l, g1, g2) * 0.75))) 
                 list_1 <- vector("list")
@@ -44,6 +44,9 @@ clust_it <- function(pop, m = 5, l = 4, g1 = 5, g2 = g1 * 0.5, d1 = 1000, d2 = d
                                                pop$max_n_cl >= ifelse(j <= l, g1, g2), 0, 1)
                 if (nrow(pop) == sum(pop$a_reasig) & j == 1) {
                         stop("Your g1 value is too high for your population dispersion")    
+                }
+                if (nrow(pop) == sum(pop$a_reasig) & j == m + 1) {
+                        stop("Your g2 value is too high for your population dispersion")   
                 }
                 clusterizados[[j]] <- subset(pop, pop$a_reasig == 0)
                 clusterizados[[j]]$round <- as.factor(j)