+++ NEWS.md
@@ -2,6 +2,19 @@
 
 -------------------------------------------------------
 
+## EPP v0.3.3 (2021-09-22) 
+
+  * leafepp debug in t = "exist" about center$capacity popup
+  * debug assign_clust line 15
+  * categorization of centers by coverage in leafepp(t = "proy")
+  * add assign_nn() function
+  * add addSearchOSM() in leafepp
+  * add labels_es dataset to translate leafepp maps 
+  * adaptation to osrm update
+  * add site by pkgdown
+  * add assertthat to check internet availability
+  * debug typo in assign_clust of p_weight
+
 ## EPP v0.3.2 (2021-08-03)
 
 Launch of **EPP** v0.3.2 on [GitHUb](https://github.com/RichDeto/EPP) with the following functions: